move paygate, fed and route into pkg
