+++ cmd/server/disclaimers.go
@@ -15,8 +15,8 @@ import (
 	"github.com/moov-io/base"
 	"github.com/moov-io/base/admin"
 	moovhttp "github.com/moov-io/base/http"
-	"github.com/moov-io/customers/cmd/server/route"
 	"github.com/moov-io/customers/pkg/client"
+	"github.com/moov-io/customers/pkg/route"
 
 	"github.com/go-kit/kit/log"
 	"github.com/gorilla/mux"
