+++ cmd/server/documents.go
@@ -18,8 +18,8 @@ import (
 
 	"github.com/moov-io/base"
 	moovhttp "github.com/moov-io/base/http"
-	"github.com/moov-io/customers/cmd/server/route"
 	"github.com/moov-io/customers/pkg/client"
+	"github.com/moov-io/customers/pkg/route"
 
 	"github.com/go-kit/kit/log"
 	"github.com/gorilla/mux"
