+++ cmd/server/files.go
@@ -14,7 +14,7 @@ import (
 	"time"
 
 	moovhttp "github.com/moov-io/base/http"
-	"github.com/moov-io/customers/cmd/server/route"
+	"github.com/moov-io/customers/pkg/route"
 
 	"github.com/go-kit/kit/log"
 	"github.com/gorilla/mux"
