+++ pkg/accounts/micro_deposits_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 
 	"github.com/moov-io/base"
-	"github.com/moov-io/customers/cmd/server/paygate"
+	"github.com/moov-io/customers/pkg/paygate"
 	"github.com/moov-io/paygate/pkg/client"
 )
