+++ pkg/accounts/router_test.go
@@ -23,8 +23,8 @@ import (
 	"github.com/moov-io/base"
 	"github.com/stretchr/testify/require"
 
-	"github.com/moov-io/customers/cmd/server/fed"
 	"github.com/moov-io/customers/pkg/client"
+	"github.com/moov-io/customers/pkg/fed"
 	"github.com/moov-io/customers/pkg/secrets"
 
 	"github.com/moov-io/customers/pkg/validator"
